+++ CHANGELOG.md
@@ -3,15 +3,22 @@
 All notable changes to this project will be documented in this file.
 
 
+## [1.3.0](https://github.com/cperuffo3/wedding-thank-you-card-generator/compare/v1.2.0...v1.3.0) (2026-01-02)
+
+### Features
+
+* improve copy ux ([a27ea21](https://github.com/cperuffo3/wedding-thank-you-card-generator/commit/a27ea2123575777e9d14c20a7bd7433f7e71adb3))
+* improve copy ux for cards ([ee27878](https://github.com/cperuffo3/wedding-thank-you-card-generator/commit/ee27878cf6e4b1652f344d3564d6da33b7578ee6))
+
 ## [1.2.0](https://github.com/cperuffo3/wedding-thank-you-card-generator/compare/v1.1.1...v1.2.0) (2026-01-01)
 
 ### Features
 
-* warning dialog on close to prevent unsaved work ([f4e2b58](https://github.com/cperuffo3/wedding-thank-you-card-generator/commit/f4e2b587ef1e7b0341a3f1c7b9f76dc498daafdd))
+- warning dialog on close to prevent unsaved work ([f4e2b58](https://github.com/cperuffo3/wedding-thank-you-card-generator/commit/f4e2b587ef1e7b0341a3f1c7b9f76dc498daafdd))
 
 ### Bug Fixes
 
-* layout height issues on editor content pt 2 ([e8e0567](https://github.com/cperuffo3/wedding-thank-you-card-generator/commit/e8e05678f1233d456cbd4298a599151ec212ba01))
+- layout height issues on editor content pt 2 ([e8e0567](https://github.com/cperuffo3/wedding-thank-you-card-generator/commit/e8e05678f1233d456cbd4298a599151ec212ba01))
 
 ## [1.1.1](https://github.com/cperuffo3/wedding-thank-you-card-generator/compare/v1.1.0...v1.1.1) (2026-01-01)